pyproject.toml
[project]
name = "langgraph_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.9.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.66",
]
